requirements.txt
dash==2.14.2
dash-html-components==2.0.0
dash-core-components==2.0.0
plotly==5.18.0
pandas==2.1.4
azure-storage-blob==12.19.0
python-dotenv==1.0.0
scikit-learn==1.3.1
pytest==7.4.4
gunicorn==21.2.0
memory-profiler==0.61.0